refactor: add ts to redux middleware, move searchVideos to videoActions.ts
