+++ src/types/SearchStore.ts
@@ -1,4 +1,4 @@
 export interface SearchStore {
   keyword: string;
-  nextPageToken?: string;
+  nextPageToken: string;
 }
